Rename integration-tests/infinispan-cache-jpa to integration-tests/hibernate-orm-cache
Because this module has NOTHING to do with Infinispan, it relies on a Caffeine-based cache.
