+++ .github/quarkus-github-bot.yml
@@ -153,7 +153,7 @@ triage:
         - integration-tests/jpa
         - integration-tests/hibernate-envers
         - integration-tests/hibernate-orm-tenancy
-        - integration-tests/infinispan-cache-jpa
+        - integration-tests/hibernate-orm-cache
     - id: hibernate-search
       labels: [area/hibernate-search]
       expression: |
@@ -401,7 +401,7 @@ triage:
       notify: [karesti, wburns]
       directories:
         - extensions/infinispan-client/
-        - integration-tests/infinispan-cache-jpa/
+        - integration-tests/infinispan-cache/
         - integration-tests/infinispan-client/
     - id: grpc
       labels: [area/grpc]
