+++ integration-tests/hibernate-orm-cache/src/main/resources/application.properties
