+++ integration-tests/hibernate-orm-cache/src/test/java/io/quarkus/it/hibernate/orm/cache/HibernateOrmCacheFunctionalityInGraalITCase.java
@@ -1,0 +1,8 @@
+package io.quarkus.it.hibernate.orm.cache;
+
+import io.quarkus.test.junit.QuarkusIntegrationTest;
+
+@QuarkusIntegrationTest
+public class HibernateOrmCacheFunctionalityInGraalITCase extends HibernateOrmCacheFunctionalityTest {
+
+}
