--- integration-tests/infinispan-cache-jpa/src/test/java/io/quarkus/it/infinispan/cache/jpa/InfinispanCacheJPAFunctionalityInGraalITCase.java
@@ -1,11 +1,0 @@
-package io.quarkus.it.infinispan.cache.jpa;
-
-import io.quarkus.test.junit.QuarkusIntegrationTest;
-
-/**
- * Test various JPA operations running in native mode
- */
-@QuarkusIntegrationTest
-public class InfinispanCacheJPAFunctionalityInGraalITCase extends InfinispanCacheJPAFunctionalityTest {
-
-}
